🔨 Refactor: 클래스 리팩토링
